Adding accessors and error catching
Adding in some accessor methods to carto, allowing you to get a
machine's info, as well as all machines in a room or on a floor.

Also adding in some simple error catching, so if you try to get a
machine that doesn't exist, the program will exit. Very basic at the
moment.